convert React Native app to typescript
