+++ PhotoFrame/App.js
@@ -1,50 +1,2 @@
-import React from "react";
-import { Alert, Button, StyleSheet, Text, View } from "react-native";
-import { Facebook } from "expo";
-
-const FACEBOOK_ID = "301045530751709";
-
-export default class App extends React.Component {
-  logIn = async () => {
-    try {
-      const {
-        type,
-        token,
-        expires,
-        permissions,
-        declinedPermissions
-      } = await Facebook.logInWithReadPermissionsAsync(FACEBOOK_ID, {
-        permissions: ["public_profile", "instagram_basic"]
-      });
-      if (type === "success") {
-        // Get the user's name using Facebook's Graph API
-        const response = await fetch(
-          `https://graph.facebook.com/me?access_token=${token}`
-        );
-        Alert.alert("Logged in!", `Hi ${(await response.json()).name}!`);
-      } else {
-        // type === 'cancel'
-      }
-    } catch ({ message }) {
-      alert(`Facebook Login Error: ${message}`);
-    }
-  };
-
-  render() {
-    return (
-      <View style={styles.container}>
-        <Text>Open up App.js to start working on your app!</Text>
-        <Button onPress={this.logIn} title="Log in with Facebook" />
-      </View>
-    );
-  }
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-    alignItems: "center",
-    justifyContent: "center"
-  }
-});
+import App from "./src/App";
+export default App;
