+++ PhotoFrame/package.json
@@ -7,6 +7,8 @@
     "eject": "expo eject"
   },
   "dependencies": {
+    "@types/expo": "^31.0.5",
+    "@types/react-native": "^0.57.23",
     "expo": "^31.0.2",
     "react": "16.5.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-31.0.0.tar.gz"
